+++ packages/vue3/src/useRemember.ts
@@ -1,5 +1,4 @@
 import { router } from '@inertiajs/core'
-import cloneDeep from 'lodash.clonedeep'
 import { isReactive, reactive, ref, Ref, watch } from 'vue'
 
 export default function useRemember<T extends object>(
@@ -18,7 +17,7 @@ export default function useRemember<T extends object>(
   watch(
     remembered,
     (newValue) => {
-      router.remember(cloneDeep(hasCallbacks ? data.__remember() : newValue), key)
+      router.remember(hasCallbacks ? data.__remember() : newValue, key)
     },
     { immediate: true, deep: true },
   )